api/pyproject.toml
[tool.poetry]
name = "django-api"
version = "0.1.0"
description = ""
authors = ["mbml84 <boumax84@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.5"
djangorestframework-api-key = "^2.3.0"
pillow = "^9.4.0"
torch = "^1.13.1"
numpy = "^1.24.2"
whitenoise = "^6.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
